+++ packages/calcite-components/src/components/autocomplete/autocomplete.tsx
@@ -1,3 +1,4 @@
+// @ts-strict-ignore
 import { PropertyValues } from "lit";
 import {
   LitElement,
@@ -11,7 +12,7 @@ import {
   LuminaJsx,
 } from "@arcgis/lumina";
 import { useWatchAttributes } from "@arcgis/components-controllers";
-import { debounce } from "lodash-es";
+import { debounce, escapeRegExp } from "lodash-es";
 import {
   FlipPlacement,
   FloatingCSS,
@@ -59,7 +60,7 @@ import { Validation } from "../functional/Validation";
 import { createObserver } from "../../utils/observers";
 import { styles } from "./autocomplete.scss";
 import T9nStrings from "./assets/t9n/messages.en.json";
-import { CSS, ICONS, IDS, SLOTS } from "./resources";
+import { CSS, IDS, SLOTS } from "./resources";
 
 const groupItemSelector = "calcite-autocomplete-item-group";
 const itemSelector = "calcite-autocomplete-item";
@@ -129,6 +130,8 @@ export class Autocomplete
   referenceEl: Input["el"];
 
   transitionEl: HTMLDivElement;
+
+  private inputValueMatchPattern: RegExp;
 
   // #endregion
 
@@ -341,6 +344,23 @@ export class Autocomplete
   }
 
   /**
+   * Scrolls the component's content to a specified set of coordinates.
+   *
+   * @example
+   * myAutocomplete.scrollContentTo({
+   *   left: 0, // Specifies the number of pixels along the X axis to scroll the window or element.
+   *   top: 0, // Specifies the number of pixels along the Y axis to scroll the window or element
+   *   behavior: "auto" // Specifies whether the scrolling should animate smoothly (smooth), or happen instantly in a single jump (auto, the default value).
+   * });
+   * @param options - allows specific coordinates to be defined.
+   * @returns - promise that resolves once the content is scrolled to.
+   */
+  @method()
+  async scrollContentTo(options?: ScrollToOptions): Promise<void> {
+    this.transitionEl?.scrollTo(options);
+  }
+
+  /**
    * Selects the text of the component's `value`.
    *
    * @returns {Promise<void>}
@@ -444,12 +464,27 @@ export class Autocomplete
       this.reposition(true);
     }
 
-    if (changes.has("scale") && (this.hasUpdated || this.scale !== "m")) {
+    let itemsAndGroupsUpdated = false;
+
+    if (changes.has("inputValue") && (this.hasUpdated || this.inputValue)) {
+      this.inputValueMatchPattern =
+        this.inputValue && new RegExp(`(${escapeRegExp(this.inputValue)})`, "i");
       this.updateItems();
       this.updateGroups();
+      itemsAndGroupsUpdated = true;
     }
 
-    if (changes.has("activeIndex") && (this.hasUpdated || this.activeIndex !== -1)) {
+    if (!itemsAndGroupsUpdated && changes.has("scale") && (this.hasUpdated || this.scale !== "m")) {
+      this.updateItems();
+      this.updateGroups();
+      itemsAndGroupsUpdated = true;
+    }
+
+    if (
+      !itemsAndGroupsUpdated &&
+      changes.has("activeIndex") &&
+      (this.hasUpdated || this.activeIndex !== -1)
+    ) {
       this.updateItems();
     }
   }
@@ -582,7 +617,12 @@ export class Autocomplete
   private updateItems(): void {
     let activeDescendant: string = null;
 
-    this.items.forEach((item, index) => {
+    this.items.forEach((item) => {
+      item.scale = this.scale;
+      item.inputValueMatchPattern = this.inputValueMatchPattern;
+    });
+
+    this.enabledItems.forEach((item, index) => {
       const isActive = index === this.activeIndex;
 
       if (isActive) {
@@ -590,7 +630,6 @@ export class Autocomplete
       }
 
       item.active = isActive;
-      item.scale = this.scale;
     });
 
     this.activeDescendant = activeDescendant;
@@ -618,12 +657,6 @@ export class Autocomplete
     this.updateGroups();
   }
 
-  private getIcon(): IconNameOrString {
-    const { icon } = this;
-
-    return icon === true ? ICONS.search : icon || ICONS.search;
-  }
-
   private setReferenceEl(el: Input["el"]): void {
     this.referenceEl = el;
 
@@ -645,6 +678,8 @@ export class Autocomplete
 
     const { open, activeIndex, enabledItems } = this;
 
+    const activeItem = enabledItems.length && activeIndex > -1 ? enabledItems[activeIndex] : null;
+
     switch (key) {
       case "Escape":
         if (open) {
@@ -656,8 +691,8 @@ export class Autocomplete
         this.open = false;
         break;
       case "Enter":
-        if (open && activeIndex > -1) {
-          this.value = enabledItems[activeIndex].value;
+        if (open && activeItem) {
+          this.value = activeItem.value;
           this.emitChange();
           this.open = false;
           event.preventDefault();
@@ -668,28 +703,44 @@ export class Autocomplete
         }
         break;
       case "ArrowDown":
-        this.open = true;
-        this.activeIndex =
-          activeIndex !== -1 ? Math.min(activeIndex + 1, enabledItems.length - 1) : 0;
-        event.preventDefault();
+        if (enabledItems.length) {
+          this.open = true;
+          this.activeIndex =
+            activeIndex !== -1 ? Math.min(activeIndex + 1, enabledItems.length - 1) : 0;
+          this.scrollToActiveItem();
+          event.preventDefault();
+        }
         break;
       case "ArrowUp":
-        this.open = true;
-        this.activeIndex =
-          activeIndex !== -1 ? Math.max(activeIndex - 1, 0) : enabledItems.length - 1;
-        event.preventDefault();
+        if (enabledItems.length) {
+          this.open = true;
+          this.activeIndex =
+            activeIndex !== -1 ? Math.max(activeIndex - 1, 0) : enabledItems.length - 1;
+          this.scrollToActiveItem();
+          event.preventDefault();
+        }
         break;
       case "Home":
-        this.open = true;
-        this.activeIndex = 0;
-        event.preventDefault();
+        if (enabledItems.length) {
+          this.open = true;
+          this.activeIndex = 0;
+          this.scrollToActiveItem();
+          event.preventDefault();
+        }
         break;
       case "End":
-        this.open = true;
-        this.activeIndex = enabledItems.length - 1;
-        event.preventDefault();
+        if (enabledItems.length) {
+          this.open = true;
+          this.activeIndex = enabledItems.length - 1;
+          this.scrollToActiveItem();
+          event.preventDefault();
+        }
         break;
     }
+  }
+
+  private scrollToActiveItem(): void {
+    this.enabledItems[this.activeIndex]?.scrollIntoView({ block: "nearest" });
   }
 
   private changeHandler(event: CustomEvent): void {
@@ -750,7 +801,7 @@ export class Autocomplete
             disabled={disabled}
             enterKeyHint={enterKeyHint}
             form={this.form}
-            icon={this.getIcon()}
+            icon={this.icon ?? true}
             iconFlipRtl={this.iconFlipRtl}
             id={inputId}
             inputMode={inputMode}