+++ packages/calcite-components/src/components/tab-nav/tab-nav.e2e.ts
@@ -1,3 +1,4 @@
+// @ts-strict-ignore
 import { newE2EPage, E2EPage, E2EElement } from "@arcgis/lumina-compiler/puppeteerTesting";
 import { describe, expect, it, beforeEach } from "vitest";
 import { accessible, defaults, hidden, renders, t9n } from "../../tests/commonTests";