+++ packages/calcite-components/src/components/text-area/text-area.tsx
@@ -1,3 +1,4 @@
+// @ts-strict-ignore
 import { throttle } from "lodash-es";
 import { createRef } from "lit-html/directives/ref.js";
 import {
@@ -204,7 +205,6 @@ export class TextArea
   /**
    * When `true`, the component's `value` can be read, but cannot be modified.
    *
-   * @readonly
    * @mdn [readOnly](https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/readonly)
    */
   @property({ reflect: true }) readOnly = false;