+++ packages/calcite-components/CHANGELOG.md
@@ -3,6 +3,100 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.0-next.92](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@3.0.0-next.91...@esri/calcite-components@3.0.0-next.92) (2025-01-08)
+
+### Features
+
+- **menu-item:** add component tokens ([#10654](https://github.com/Esri/calcite-design-system/issues/10654)) ([9dc1262](https://github.com/Esri/calcite-design-system/commit/9dc12626331dbf9252ee4f15fc42dab3358029f1)), closes [#7180](https://github.com/Esri/calcite-design-system/issues/7180)
+- **rating:** add component tokens ([#11150](https://github.com/Esri/calcite-design-system/issues/11150)) ([6dbb559](https://github.com/Esri/calcite-design-system/commit/6dbb55946597559e37ce7c1747386a3a69533025)), closes [#7180](https://github.com/Esri/calcite-design-system/issues/7180)
+
+### Bug Fixes
+
+- **docs:** add docs package export ([#11221](https://github.com/Esri/calcite-design-system/issues/11221)) ([4b0d804](https://github.com/Esri/calcite-design-system/commit/4b0d8043ddcd2d05451253256a6812d5427a69eb)), closes [#10731](https://github.com/Esri/calcite-design-system/issues/10731)
+- **docs:** use updated translation bundle paths to generate `translations.json` ([#11219](https://github.com/Esri/calcite-design-system/issues/11219)) ([dce5635](https://github.com/Esri/calcite-design-system/commit/dce56356568f606004f9210ab21cbbcf7dcb94af)), closes [#10731](https://github.com/Esri/calcite-design-system/issues/10731)
+- **tile:** wrap overflowing text content ([#11213](https://github.com/Esri/calcite-design-system/issues/11213)) ([732933a](https://github.com/Esri/calcite-design-system/commit/732933a8ed6e0da21edef9d6b8d6b59b363d7292)), closes [#11170](https://github.com/Esri/calcite-design-system/issues/11170)
+
+## [3.0.0-next.91](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@3.0.0-next.90...@esri/calcite-components@3.0.0-next.91) (2025-01-07)
+
+### Bug Fixes
+
+- **text-area:** make `readOnly` prop writable ([#11218](https://github.com/Esri/calcite-design-system/issues/11218)) ([0cbf766](https://github.com/Esri/calcite-design-system/commit/0cbf766156e3b0ec7a662571606af62895e23f59)), closes [#11217](https://github.com/Esri/calcite-design-system/issues/11217) [#10731](https://github.com/Esri/calcite-design-system/issues/10731) [#10310](https://github.com/Esri/calcite-design-system/issues/10310)
+
+## [3.0.0-next.90](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@3.0.0-next.89...@esri/calcite-components@3.0.0-next.90) (2025-01-06)
+
+### Features
+
+- **notice:** add component tokens ([#11042](https://github.com/Esri/calcite-design-system/issues/11042)) ([a61c15a](https://github.com/Esri/calcite-design-system/commit/a61c15af46e772956a58fcc71f873b910897097e)), closes [#7180](https://github.com/Esri/calcite-design-system/issues/7180)
+
+### Bug Fixes
+
+- **input:** correctly apply placeholder styles ([#11107](https://github.com/Esri/calcite-design-system/issues/11107)) ([7fa1eea](https://github.com/Esri/calcite-design-system/commit/7fa1eeaeaf090b1cf4e835a6e3bcd678c3918173)), closes [#8967](https://github.com/Esri/calcite-design-system/issues/8967)
+- **time-picker:** display correct localized hour based on hour-format when no value is set ([#11200](https://github.com/Esri/calcite-design-system/issues/11200)) ([a938192](https://github.com/Esri/calcite-design-system/commit/a938192b510a5c57fa7613bbea7e965d540243c7)), closes [#11198](https://github.com/Esri/calcite-design-system/issues/11198)
+
+## [3.0.0-next.89](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@3.0.0-next.88...@esri/calcite-components@3.0.0-next.89) (2025-01-04)
+
+### Features
+
+- **combobox, input-time-zone:** highlight text matches consistently ([#11193](https://github.com/Esri/calcite-design-system/issues/11193)) ([75f54de](https://github.com/Esri/calcite-design-system/commit/75f54de466fa2938af9a333cb9c3e4b8ed00460b)), closes [#11173](https://github.com/Esri/calcite-design-system/issues/11173)
+
+## [3.0.0-next.88](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@3.0.0-next.87...@esri/calcite-components@3.0.0-next.88) (2025-01-03)
+
+### Features
+
+- **combobox:** add `filterProps` ([#11180](https://github.com/Esri/calcite-design-system/issues/11180)) ([09d5a27](https://github.com/Esri/calcite-design-system/commit/09d5a2777ebe3c915a3284efca02768cf5c2823a)), closes [#8043](https://github.com/Esri/calcite-design-system/issues/8043)
+
+### Bug Fixes
+
+- **autocomplete:** fix item key functionality ([#11191](https://github.com/Esri/calcite-design-system/issues/11191)) ([d2ced69](https://github.com/Esri/calcite-design-system/commit/d2ced69f4967f17bddb39a2a9e0256ec59c5a0bf)), closes [#11154](https://github.com/Esri/calcite-design-system/issues/11154)
+- **list:** include groups in filtering ([#10664](https://github.com/Esri/calcite-design-system/issues/10664)) ([904623b](https://github.com/Esri/calcite-design-system/commit/904623be3cff426e71d51001a29df188a72ea00b)), closes [#7702](https://github.com/Esri/calcite-design-system/issues/7702)
+
+## [3.0.0-next.87](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@3.0.0-next.86...@esri/calcite-components@3.0.0-next.87) (2025-01-02)
+
+**Note:** Version bump only for package @esri/calcite-components
+
+## [3.0.0-next.86](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@3.0.0-next.85...@esri/calcite-components@3.0.0-next.86) (2025-01-01)
+
+### Features
+
+- **input-time-picker:** add hour-format property ([#10997](https://github.com/Esri/calcite-design-system/issues/10997)) ([52e9013](https://github.com/Esri/calcite-design-system/commit/52e9013883728f593942482dbade23724754824b)), closes [#4756](https://github.com/Esri/calcite-design-system/issues/4756)
+
+## [3.0.0-next.85](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@3.0.0-next.84...@esri/calcite-components@3.0.0-next.85) (2024-12-31)
+
+### Features
+
+- **autocomplete:** add scrollContentTo method and automatically scroll when navigating items via keyboard ([#11153](https://github.com/Esri/calcite-design-system/issues/11153)) ([be57a3b](https://github.com/Esri/calcite-design-system/commit/be57a3b64fabd321a06412fe8476d7b06ac84c69)), closes [#11152](https://github.com/Esri/calcite-design-system/issues/11152)
+
+## [3.0.0-next.84](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@3.0.0-next.83...@esri/calcite-components@3.0.0-next.84) (2024-12-31)
+
+### Features
+
+- **autocomplete:** highlight text matching the inputValue ([#11155](https://github.com/Esri/calcite-design-system/issues/11155)) ([a43982b](https://github.com/Esri/calcite-design-system/commit/a43982b2798fcfc6dd96aa98b668abd008e86626)), closes [#11154](https://github.com/Esri/calcite-design-system/issues/11154)
+
+## [3.0.0-next.83](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@3.0.0-next.82...@esri/calcite-components@3.0.0-next.83) (2024-12-31)
+
+### Features
+
+- **block:** add component tokens ([#11014](https://github.com/Esri/calcite-design-system/issues/11014)) ([3ca4c98](https://github.com/Esri/calcite-design-system/commit/3ca4c98d8009e5f7d594bcaa1f829a035c31dbab)), closes [#7180](https://github.com/Esri/calcite-design-system/issues/7180)
+
+### Bug Fixes
+
+- **input-time-zone:** keep selection in sync when item-related props are set along with value ([#11166](https://github.com/Esri/calcite-design-system/issues/11166)) ([47c3e9b](https://github.com/Esri/calcite-design-system/commit/47c3e9b38ff5e131d8bca6b1bb7ecb74182fbd0f)), closes [#11128](https://github.com/Esri/calcite-design-system/issues/11128)
+- **list-item:** improve drag handle spacing ([#11156](https://github.com/Esri/calcite-design-system/issues/11156)) ([7b8b7d8](https://github.com/Esri/calcite-design-system/commit/7b8b7d89fe517800d24c3e30cd285432d5a0e3dc)), closes [#11125](https://github.com/Esri/calcite-design-system/issues/11125)
+
+## [3.0.0-next.82](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@3.0.0-next.81...@esri/calcite-components@3.0.0-next.82) (2024-12-28)
+
+### Features
+
+- **accordion-item:** update token naming schema ([#11070](https://github.com/Esri/calcite-design-system/issues/11070)) ([d624985](https://github.com/Esri/calcite-design-system/commit/d624985323637d7a04d506b8ffc7fdd6468ae03a)), closes [#10894](https://github.com/Esri/calcite-design-system/issues/10894)
+- **block-section:** add component tokens ([#11032](https://github.com/Esri/calcite-design-system/issues/11032)) ([1f9449b](https://github.com/Esri/calcite-design-system/commit/1f9449b03af83e334e902d014af82d3f286512bf)), closes [#7180](https://github.com/Esri/calcite-design-system/issues/7180)
+
+### Bug Fixes
+
+- **action-menu:** fix toggle logic when `action-menu` is reconnected ([#11139](https://github.com/Esri/calcite-design-system/issues/11139)) ([5874d6f](https://github.com/Esri/calcite-design-system/commit/5874d6f7cf1670d57dd4a884a2a8375a727aa153)), closes [#10731](https://github.com/Esri/calcite-design-system/issues/10731) [#10310](https://github.com/Esri/calcite-design-system/issues/10310)
+- **flow:** set focus on the selected flow item ([#11127](https://github.com/Esri/calcite-design-system/issues/11127)) ([7924c7b](https://github.com/Esri/calcite-design-system/commit/7924c7bdf8f5dc99f03a255d58eec109ddc4e247)), closes [#5072](https://github.com/Esri/calcite-design-system/issues/5072)
+- **list-item:** ensure consistent focus border color by referencing updated tokens ([#11130](https://github.com/Esri/calcite-design-system/issues/11130)) ([d1f6f1d](https://github.com/Esri/calcite-design-system/commit/d1f6f1dd6181d55eb91971cd0ccd33ce5a96fc61))
+- **popover, tooltip:** skip ref setup logic on component removal ([#11132](https://github.com/Esri/calcite-design-system/issues/11132)) ([e67b1fe](https://github.com/Esri/calcite-design-system/commit/e67b1fe9c95218611edd37a51aa7fe3e9d70d410)), closes [#10731](https://github.com/Esri/calcite-design-system/issues/10731) [#10310](https://github.com/Esri/calcite-design-system/issues/10310)
+
 ## [3.0.0-next.81](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@3.0.0-next.80...@esri/calcite-components@3.0.0-next.81) (2024-12-23)
 
 ### Features